+++ packages/tokens/src/story-ui/StackList.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react';
-import styles from './story.module.css';
+import styles from './story.module.scss';
 
 interface StackListProps {
   title: string;
