+++ packages/tokens/src/types/css.d.ts
@@ -1,4 +1,4 @@
-declare module '*.module.css' {
+declare module '*.module.scss' {
   const classes: { [key: string]: string };
   export default classes;
 }
